fix(arena/web): hydrate Previous Runs from every JSON file, not index.json
LoadResultsIntoStore relied on JSONResultRepository.LoadResults, which
reads run_ids out of index.json. index.json is overwritten by each
CLI invocation, so on server boot only the latest batch's run IDs
were hydrated into the in-memory store — even though every prior run
sat on disk as <runID>.json.

Scan the output directory directly. Web-triggered runs that bypass
index.json are now visible after a server restart, and the homepage's
Previous Runs reflects the full history on initial load.
